added beginning of edits to student and moved css files to new css folder
