--- src/Four0Four.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Four0Four = () => {
-
-    return (
-        <>
-            <br />
-            <br />
-            <h1>SORRY, PAGE NOT FOUND ......</h1>
-        </>
-    )
-
-}
-export default Four0Four
