+++ src/css/students.css
@@ -7,3 +7,27 @@
     font-size: x-large;
     background-color: rgba(238, 238, 247, 0.918);
 }
+
+.students{
+    text-shadow: .02vh .03vh #717375;
+   
+}
+.students-blur {
+    background-color: black;
+}
+
+
+.new-student-form{
+    margin-left: 35%;
+    width: 35%;
+  }
+
+  .student-input{
+    /* display: flex; */
+    /* justify-items: center;
+    align-items: center;
+    margin: 5vh;
+    padding: 1vw;
+    width: auto;
+    background-color: rgb(244, 245, 248); */
+  }
